pyproject.toml
[project]
name = "tbxt-hackathon"
version = "0.1.0"
description = "TBXT hit identification hackathon — ligand-based virtual screening with structure-based validation"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Nate Harms", email = "nate@harmsinformatics.org"},
    {name = "Raymond Gasper"},
]
requires-python = ">=3.11"
dependencies = [
    # cheminformatics
    "rdkit>=2024.3.1",
    # structural biology
    "biopython>=1.84",
    "numpy>=1.26.0",
    # data
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "msoffcrypto-tool>=5.4.0",
    # ML
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    # notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    # viz
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    # docking (vina installed via conda: conda install -c conda-forge vina)
    "meeko>=0.6",
    "gemmi>=0.7",
    "py3Dmol>=2.0.0",
    # utilities
    "tqdm>=4.66.0",
    "chemprop>=2.2.0",
    "pacmap>=0.7",
    "polars>=1.20",
    "dataframely>=1.2",
    "loguru>=0.7",
    "typeguard>=4",
    "joblib>=1.4",
    "statsmodels>=0.14",
    "duckdb",
    "pyarrow",
    "optuna>=4.8.0",
    "umap-learn>=0.5.12",
    "onepot>=0.1.6",
    "apple-smi>=0.1.4",
]

[dependency-groups]
dev = [
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tbxt_hackathon"]

